Reorginized the code a bit
